feat: add-new-component-refactoring
