+++ src/components/index.js
@@ -2,3 +2,4 @@ export { default as Home } from './Home';
 export { default as Signup } from './Signup';
 export { default as Slider } from './Home/Slider';
 export { default as Signin } from './Signin';
+export { default as Header } from './Home/Header';
